requirements.txt
flask>=2.0.0
pandas>=1.3.0
python-miio>=0.5.12
paho-mqtt>=1.6.1
plotly>=5.13.0
python-dotenv>=0.19.0 